Move confirm-transaction/conf-tx.js to confirm-signature-request
