+++ development/ts-migration-dashboard/files-to-convert.json
@@ -1241,6 +1241,7 @@
   "ui/pages/confirm-send-token/confirm-send-token.js",
   "ui/pages/confirm-send-token/confirm-send-token.stories.js",
   "ui/pages/confirm-send-token/index.js",
+  "ui/pages/confirm-signature-request/index.js",
   "ui/pages/confirm-token-transaction-base/confirm-token-transaction-base.js",
   "ui/pages/confirm-token-transaction-base/confirm-token-transaction-base.stories.js",
   "ui/pages/confirm-token-transaction-base/index.js",
@@ -1254,7 +1255,6 @@
   "ui/pages/confirm-transaction-switch/confirm-transaction-switch.component.js",
   "ui/pages/confirm-transaction-switch/confirm-transaction-switch.container.js",
   "ui/pages/confirm-transaction-switch/index.js",
-  "ui/pages/confirm-transaction/conf-tx.js",
   "ui/pages/confirm-transaction/confirm-token-transaction-switch.js",
   "ui/pages/confirm-transaction/confirm-transaction.component.js",
   "ui/pages/confirm-transaction/confirm-transaction.container.js",
