+++ ui/pages/confirm-signature-request/index.test.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import configureMockStore from 'redux-mock-store';
 import { renderWithProvider } from '../../../test/lib/render-helpers';
-import ConfTx from './conf-tx';
+import ConfTx from '.';
 
 const mockState = {
   metamask: {
